onnx_server/requirements.txt
Flask
onnxruntime
Pillow
numpy
onnx

